requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
--trusted-host pypi.python.org
--trusted-host pypi.org
--trusted-host files.pythonhosted.org

azure-common==1.1.28
    # via azure-mgmt-compute
azure-core==1.32.0
    # via
    #   azure-identity
    #   azure-mgmt-core
    #   azure-storage-blob
azure-functions==1.21.3
    # via -r requirements.in
azure-identity==1.19.0
    # via -r requirements.in
azure-mgmt-compute==33.1.0
    # via -r requirements.in
azure-mgmt-core==1.5.0
    # via azure-mgmt-compute
azure-storage-blob==12.24.0
    # via -r requirements.in
certifi==2024.12.14
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
cryptography==44.0.0
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
idna==3.10
    # via requests
isodate==0.7.2
    # via
    #   azure-mgmt-compute
    #   azure-storage-blob
msal==1.31.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.2.0
    # via azure-identity
portalocker==2.10.1
    # via msal-extensions
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.10.1
    # via
    #   msal
    #   pyjwt
requests==2.32.3
    # via
    #   -r requirements.in
    #   azure-core
    #   msal
semver==3.0.2
    # via -r requirements.in
six==1.17.0
    # via azure-core
typing-extensions==4.12.2
    # via
    #   azure-core
    #   azure-identity
    #   azure-mgmt-compute
    #   azure-storage-blob
urllib3==2.3.0
    # via requests
